report_app/requirements.txt
fpdf2
matplotlib